set non-zero platform fee

+++ auction-server/src/auction/service/verification.rs
@@ -602,11 +602,8 @@ impl Service<Svm> {
                     ),
                 };
 
-                let permission_account = get_quote_permission_key(
-                    &tokens,
-                    &user_wallet,
-                    swap_data.referral_fee_bps as u16,
-                );
+                let permission_account =
+                    get_quote_permission_key(&tokens, &user_wallet, swap_data.referral_fee_bps);
 
                 Ok(BidDataSvm {
                     amount: bid_amount,
